update(sm): reformat bitacora, misc resources and header added
