+++ templates/resources/methods/Functions_Aux_Login.py
@@ -51,7 +51,7 @@ def verify_employee_id(token_data: dict, emp_id: int) -> bool:
 
 
 def verify_token(
-    token: str, department: str = None, emp_id: int = None
+    token: str, department: str | list = None, emp_id: int = None
 ) -> tuple[bool, dict]:
     """
     Verifies the token.
@@ -66,8 +66,14 @@ def verify_token(
             if not verify_employee_id(data, emp_id):
                 return False, {}
         elif department:
-            if not verify_department_permission(data, department):
+            if isinstance(department, list):
+                for dep in department:
+                    if verify_department_permission(data, dep):
+                        return True, data
                 return False, {}
+            elif isinstance(department, str):
+                if not verify_department_permission(data, department):
+                    return False, {}
         return True, data
     except Exception as e:
         print("errort at unpack token: ", e)
