+++ templates/resources/midleware/Functions_midleware_misc.py
@@ -84,103 +84,6 @@ def get_response_AV(
     return files_av, res, id_chat
 
 
-def handle_comment_extra(type_extra: int, comment: str, hour_in, hour_out):
-    comment_dict = split_commment(comment)
-    places = comment_dict["place"].split("<**>")
-    if len(places) > 1:
-        places[type_extra] = " "
-    comment_dict["place"] = "<**>".join(places)
-    comment_dict["times"] = "<**>".join([hour_in, hour_out])
-    comment_dict["aproved"] = 0
-    comment_out = unify_comment_dict(comment_dict)
-    return comment_out
-
-
-def get_events_from_extraordinary_sources(hour_in: str, hour_out: str, data: dict):
-    events = []
-    data_events = []
-    normal_hour_in = datetime.strptime(hour_in.split("-->")[0], "%H:%M")
-    normal_hour_out = datetime.strptime(hour_out.split("-->")[0], "%H:%M")
-    tmp_hour_in = datetime.strptime(hour_in.split("-->")[2], "%H:%M")
-    tmp_hour_out = datetime.strptime(hour_out.split("-->")[2], "%H:%M")
-    # timestamp limit hours in and out
-    tmp_limit_hour_in = datetime.strptime(hour_in.split("-->")[1], "%H:%M")
-    tmp_limit_hour_out = datetime.strptime(hour_out.split("-->")[1], "%H:%M")
-    if (
-        normal_hour_in <= tmp_hour_in <= tmp_limit_hour_in
-        and normal_hour_out <= tmp_hour_out <= tmp_limit_hour_out
-    ):
-        # print(
-        #     f"event: normal, date: {data['date']}, value: {1.0}, comment: {data['comment']}"
-        # )
-        events.append("normal")
-        data_events.append([data["date"], 1.0, data["comment"], data["contract"]])
-    else:
-        if tmp_hour_in > tmp_limit_hour_in:
-            hours_late = tmp_hour_in - tmp_limit_hour_in
-            hours_late = hours_late.total_seconds() / 3600.0
-            # print(f"event: late, date: {data['date']}, value: {int(hours_late)} hours {int(hours_late % 1 * 60)} minutes, comment: {data['comment']}")
-            events.append("atraso")
-            data_events.append(
-                [data["date"], hours_late, data["comment"], data["contract"]]
-            )
-        elif tmp_hour_in < normal_hour_in:
-            hours_early = normal_hour_in - tmp_hour_in
-            hours_early = hours_early.total_seconds() / 3600.0
-            # print(f"event: extra, date: {data['date']}, value: {int(-hours_early)} hours {int(-hours_early % 1 * 60)} minutes, comment: {data['comment']}")
-            events.append("extra")
-            data_events.append(
-                [
-                    data["date"],
-                    hours_early,
-                    handle_comment_extra(
-                        1,
-                        data["comment"],
-                        hour_in.split("-->")[2],
-                        hour_out.split("-->")[2],
-                    ),
-                    data["contract"],
-                ]
-            )
-        if tmp_hour_out > tmp_limit_hour_out:
-            hours_extra = tmp_hour_out - tmp_limit_hour_out
-            hours_extra = hours_extra.total_seconds() / 3600.0
-            # print(f"event: extra, date: {data['date']}, value: {int(hours_extra)} hours {int(hours_extra % 1 * 60)} minutes, comment: {data['comment']}")
-            events.append("extra")
-            data_events.append(
-                [
-                    data["date"],
-                    hours_extra,
-                    handle_comment_extra(
-                        0,
-                        data["comment"],
-                        hour_in.split("-->")[2],
-                        hour_out.split("-->")[2],
-                    ),
-                    data["contract"],
-                ]
-            )
-        # early leaving
-        elif tmp_hour_out < normal_hour_out:
-            early_hours = normal_hour_out - tmp_hour_out
-            early_hours = early_hours.total_seconds() / 3600.0
-            # print(f"event: early, date: {data['date']}, value: {int(-early_hours)} hours {int(-early_hours % 1 * 60)} minutes, comment: {data['comment']}")
-            events.append("early")
-            data_events.append(
-                [data["date"], early_hours, data["comment"], data["contract"]]
-            )
-    events_final = []
-    data_events_final = []
-    for event, data_event in zip(events, data_events):
-        if event in events_final:
-            index = events_final.index(event)
-            data_events_final[index][1] += data_event[1]
-        else:
-            events_final.append(event)
-            data_events_final.append(data_event)
-    return events_final, data_events_final
-
-
 def get_task_by_id_employee(id_emp: int):
     flag, error, result = get_task_by_id_emp(id_emp)
     if flag:
